Move views to their own file
