+++ app/src/main/kotlin/ar/pelotude/ohhsugoi/MangaKog.kt
@@ -12,13 +12,11 @@ import com.kotlindiscord.kord.extensions.types.respondingPaginator
 import dev.kord.common.Color
 import dev.kord.common.entity.Snowflake
 import dev.kord.core.entity.Attachment
-import dev.kord.rest.builder.interaction.*
 import dev.kord.rest.builder.message.EmbedBuilder
 import dev.kord.rest.builder.message.create.embed
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
 import java.net.URL
-import java.util.*
 
 class MangaExtension: Extension(), KoinComponent {
     private val db: MangaDatabase by inject()
@@ -212,57 +210,4 @@ class MangaExtension: Extension(), KoinComponent {
             }
         }
     }
-}
-
-
-/**
- * Builds a manga view from the scope of an [EmbedBuilder] and returns it.
- *
- * @return The [EmbedBuilder] this function was called from.
- */
-fun EmbedBuilder.mangaView(manga: MangaWithTags): EmbedBuilder {
-    title = manga.title
-
-    manga.imgURLSource?.let { imgURL ->
-        thumbnail {
-            url = imgURL.toString()
-        }
-    }
-
-    field {
-        name = "Categorías"
-
-        val tags = manga.tags.takeIf(Set<*>::isNotEmpty)?.joinToString(
-                separator = ", ",
-                prefix = " ",
-                postfix = ".".replaceFirstChar { if (it.isLowerCase()) it.titlecase(Locale.getDefault()) else it.toString() }
-        ) ?: ""
-        value = "${manga.demographic}.$tags"
-    }
-
-    field {
-        name = "Tomos"
-        value = manga.volumes?.toString() ?: "?"
-        inline = true
-    }
-    field {
-        name = "Páginas por tomo"
-        value = manga.pagesPerVolume?.let { "~$it" } ?: "?"
-        inline = true
-    }
-    field {
-        name = "Páginas por capítulo"
-        value = manga.pagesPerChapter?.let { "~$it" } ?: "?"
-        inline = true
-    }
-
-    description =
-            manga.description +
-                    "\n[Leer](${manga.link})"
-
-    footer {
-        text = if (manga.read) "Leído por el club." else "No leído por el club."
-    }
-
-    return this
 }
